flask/requirements.txt
#cassandra-driver==2.6.0
#Flask==0.10.1
#futures==3.0.3
#itsdangerous==0.24
#Jinja2==2.7.3
#MarkupSafe==0.23
#six==1.9.0
#Werkzeug==0.10.4

flask_user 
flask_sqlalchemy
flask_wtf
wtforms
flask_bcrypt
flask_login
cassandra-driver==3.24.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SocketIO==4.3.1
geomet==0.2.1.post1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
pusher==3.0.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==19.1.0
python-engineio==3.13.2
python-socketio==4.6.0
redis==3.5.3
requests==2.24.0
six==1.15.0
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
yapf==0.30.0